Move annotation utilities to API
For example, for parameter resolvers
