+++ src/main/kotlin/io/github/freya022/botcommands/internal/utils/FunctionFilter.kt
@@ -1,5 +1,6 @@
 package io.github.freya022.botcommands.internal.utils
 
+import io.github.freya022.botcommands.api.core.utils.hasAnnotationRecursive
 import io.github.freya022.botcommands.api.core.utils.isStatic
 import io.github.freya022.botcommands.api.core.utils.isSubclassOfAny
 import io.github.freya022.botcommands.api.core.utils.simpleNestedName
