+++ src/main/kotlin/io/github/freya022/botcommands/internal/utils/ReflectionUtils.kt
@@ -5,9 +5,7 @@ import io.github.freya022.botcommands.api.core.utils.isSubclassOf
 import io.github.freya022.botcommands.api.core.utils.javaMethodOrConstructor
 import io.github.freya022.botcommands.api.core.utils.simpleNestedName
 import net.dv8tion.jda.api.events.Event
-import java.lang.annotation.Inherited
 import java.lang.reflect.Method
-import java.util.*
 import java.util.concurrent.locks.ReentrantLock
 import kotlin.concurrent.withLock
 import kotlin.jvm.internal.CallableReference
@@ -203,149 +201,4 @@ internal fun KType.typeOfAtOrNullOnStar(index: Int, targetType: KClass<*>): KTyp
 internal fun <T : Any> Class<T>.safeCast(instance: Any?): T? = when {
     isInstance(instance) -> cast(instance)
     else -> null
-}
-
-internal inline fun <reified A : Annotation> KAnnotatedElement.hasAnnotationRecursive(): Boolean =
-    hasAnnotationRecursive(A::class)
-
-internal fun <A : Annotation> KAnnotatedElement.hasAnnotationRecursive(annotationType: KClass<A>): Boolean =
-    findAnnotationRecursive(annotationType) != null
-
-/**
- * Finds a single annotation of the specified type, from the annotated element.
- *
- * The search is breadth-first, and only considers annotations, not superclasses,
- * which should be handled by [Inherited].
- *
- * **Note:** only "repeatable" annotations should be inheritable
- */
-internal inline fun <reified A : Annotation> KAnnotatedElement.findAnnotationRecursive(): A? =
-    findAnnotationRecursive(A::class)
-
-/**
- * Finds a single annotation of the specified type, from the annotated element.
- *
- * The search is breadth-first, and only considers annotations, not superclasses,
- * which should be handled by [Inherited].
- *
- * **Note:** only "repeatable" annotations should be inheritable
- */
-internal fun <A : Annotation> KAnnotatedElement.findAnnotationRecursive(annotationType: KClass<A>): A? {
-    var foundAnnotation: A? = null
-    bfs(this) {
-        val annotation = annotationType.safeCast(it)
-        if (annotation != null) {
-            foundAnnotation = annotation
-            false
-        } else {
-            true
-        }
-    }
-    return foundAnnotation
-}
-
-/**
- * Finds all annotations of the specified type, from the annotated element.
- *
- * The search is breadth-first, and only considers annotations, not superclasses,
- * which should be handled by [Inherited].
- *
- * **Note:** only "repeatable" annotations should be inheritable
- */
-internal inline fun <reified A : Annotation> KAnnotatedElement.findAllAnnotations(): List<A> =
-    findAllAnnotations(A::class)
-
-/**
- * Finds all annotations of the specified type, from the annotated element.
- *
- * The search is breadth-first, and only considers annotations, not superclasses,
- * which should be handled by [Inherited].
- *
- * **Note:** only "repeatable" annotations should be inheritable
- */
-internal fun <A : Annotation> KAnnotatedElement.findAllAnnotations(annotationType: KClass<A>): List<A> = buildList {
-    bfs(this@findAllAnnotations) {
-        val annotation = annotationType.safeCast(it)
-        if (annotation != null)
-            this += annotation
-        true
-    }
-}
-
-internal data class MetaAnnotatedClass<A : Annotation>(val metaAnnotatedElement: Annotation, val metaAnnotation: A)
-
-/**
- * Finds all annotations meta-annotated with [A], from the annotated element.
- *
- * The search is breadth-first, and only considers annotations, not superclasses,
- * which should be handled by [Inherited].
- *
- * **Note:** only "repeatable" annotations should be inheritable
- */
-internal inline fun <reified A : Annotation> KAnnotatedElement.findAllAnnotationsWith(): List<MetaAnnotatedClass<A>> =
-    findAllAnnotationsWith(A::class)
-
-/**
- * Finds all annotations meta-annotated with [annotationType], from the annotated element.
- *
- * The search is breadth-first, and only considers annotations, not superclasses,
- * which should be handled by [Inherited].
- *
- * **Note:** only "repeatable" annotations should be inheritable
- */
-@Suppress("UNCHECKED_CAST")
-internal fun <A : Annotation> KAnnotatedElement.findAllAnnotationsWith(annotationType: KClass<A>): List<MetaAnnotatedClass<A>> = buildList {
-    bfs(this@findAllAnnotationsWith) {
-        bfs(it.annotationClass) { metaAnnotation ->
-            if (metaAnnotation.annotationClass == annotationType) {
-                this += MetaAnnotatedClass(it, metaAnnotation as A)
-                false
-            } else {
-                true
-            }
-        }
-
-        true
-    }
-}
-
-/**
- * Finds all annotations from the annotated element.
- *
- * The search is breadth-first, and only considers annotations, not superclasses,
- * which should be handled by [Inherited].
- *
- * **Note:** only "repeatable" annotations should be inheritable
- */
-internal fun KAnnotatedElement.getAllAnnotations(): List<Annotation> = buildList {
-    bfs(this@getAllAnnotations) {
-        this += it
-        true
-    }
-}
-
-private fun bfs(root: KAnnotatedElement, block: (annotation: Annotation) -> Boolean) {
-    // The annotation is most likely on the root element, avoid hashSetOf/ArrayDeque allocation
-    root.annotations.forEach {
-        if (!block(it)) return
-    }
-
-    // Block hasn't returned, we need to search deeper
-    deepBfs(root, block)
-}
-
-private fun deepBfs(root: KAnnotatedElement, block: (annotation: Annotation) -> Boolean) {
-    val visited = hashSetOf<KAnnotatedElement>()
-    val toVisit = ArrayDeque<KAnnotatedElement>()
-    root.annotations.forEach { toVisit.addLast(it.annotationClass) }
-
-    while (toVisit.isNotEmpty()) {
-        val element = toVisit.removeFirst()
-        if (!visited.add(element)) continue
-
-        element.annotations.forEach {
-            toVisit.addLast(it.annotationClass)
-            if (!block(it)) return
-        }
-    }
 }
